READ THIS: Lots of solutions are missing from the "complete" set that otherwise satisfy every other condition. Whether this is due to a bug in check_duplicates.c or there are genuine missing solutions is hard to say.
Anyways, changes comprise of many bug fixes
including:
1. Completely incorrect reward function
2. Missing action mask for charges exceeding
   bounds
3. Charges not being recorded in order of
   ascending magnitude
4. C module not reading charges_sum correctly due
   to mismatched datatypes
5. Added many more potentially useful things in
   checks/
6. Implmented a half-assed way to get rid of oob
   solutions

+++ src/charge_space.py
@@ -59,9 +59,9 @@ class Charge_Space(gym.spaces.Box):
                 break
 
         curr_quad, curr_cube, curr_yukawa = anomaly_quadratic(charges), anomaly_cubic(charges), yukawa(charges_sum)
-        curr_coef = np.array([curr_quad, curr_cube, curr_yukawa])
+        curr_coef = np.array([curr_quad, curr_cube, curr_yukawa], dtype=np.int32)
 
-        return charges, charges_sum, curr_coef, np.zeros((3,))
+        return charges, charges_sum, curr_coef, np.zeros((3,), dtype=np.int32)
 
     def contains(self, sample):
         if not super().contains(sample):
